Livestream/requirements.txt
Flask==2.3.2
requests==2.31.0
flask-cors==4.0.1
mysql-connector-python==8.1.0